moved some parts of the implementation to the api
