+++ server/src/main/java/org/machinemc/server/Machine.java
@@ -245,6 +245,7 @@ public final class Machine implements Server {
         console.info("Registered " + dimensionTypeManager.getDimensions().size() + " dimension types");
 
         messenger = MessengerImpl.createDefault(this);
+        console.info("Registered " + messenger.getChatTypes().size() + " chat types");
 
         // Loading biomes json file
         biomeManager = new BiomeManagerImpl(this);
